packages/shared/pyproject.toml
[project]
name = "rag_pipeline"
version = "0.1.0"
description = "Shared RAG pipeline library - Core utilities and components"
readme = "../../README.md"
requires-python = ">=3.10"
authors = [
    { name = "Elie Bruno" },
    { name = "Andrea Trugenberger" },
    { name = "Youssef Chelaifa" }
]

dependencies = [
    # Core utilities (used everywhere)
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.2",
    "requests>=2.31.0",
    "tenacity>=9.1.2",
    # Data processing
    "pandas>=2.1.3",
    "pyarrow>=14.0.1",
    "tqdm>=4.66.1",
]

[project.optional-dependencies]
# PDF processing (for workers only) - Heavy GPU dependencies
pdf = [
    # PDF libraries
    "pypdf>=3.17.0",
    "pdf2image>=1.16.3",
    "pdfplumber>=0.10.3",
    "pillow>=10.1.0",
    "pymupdf>=1.26.6",
    "opencv-python>=4.12.0.88",
    # ML & Transformers (for Dolphin model)
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.35.0",
    "accelerate>=0.24.1",
    "safetensors>=0.4.0",
    "sentence-transformers>=2.2.2",
]

# Vector operations (for API and workers)
vector = [
    "openai>=1.3.5",
    "tiktoken>=0.5.1",
    "faiss-cpu>=1.7.4",
    "weaviate-client>=3.25.3",
]

# Cloud storage (for API and workers)
cloud = [
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "aiohttp>=3.9.0",
]

# Analysis and benchmarking
analysis = [
    "matplotlib>=3.10.7",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "plotly>=6.4.0",
    "kaleido>=1.2.0",
    "jinja2>=3.1.6",
    "nbformat>=5.10.4",
]

# CLI utilities
cli = [
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "bs4>=0.0.2",
    "nltk>=3.8.1",
    "markdown>=3.5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
